Move compare related classes to their namespace.
