+++ src/Infrastructure/Command/EventSubscriber.php
@@ -9,7 +9,8 @@ use AkeneoEtl\Domain\Load\LoadResult\Failed as LoadFailed;
 use AkeneoEtl\Domain\Resource\Resource;
 use AkeneoEtl\Domain\Transform\Event\AfterTransformEvent;
 use AkeneoEtl\Domain\Transform\TransformResult\Failed as TransformFailed;
-use AkeneoEtl\Infrastructure\Command\Compare\ConsoleTableFormatter;
+use AkeneoEtl\Infrastructure\Comparer\ConsoleTableFormatter;
+use AkeneoEtl\Infrastructure\Comparer\ResourceComparer;
 use LogicException;
 use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Input\InputInterface;
